remove index and change home to index
